+++ src/engine/ReactPlugin/EditorReactPlugin.js
@@ -1,9 +1,10 @@
-import React, { useMemo, useEffect, useCallback } from "react";
+import React, { useEffect, useCallback } from "react";
 import withAlerts from "../../decorators/withAlerts";
 import withKeyBinds from "../../decorators/withKeyBinds";
 import withMenuHandler from "../../decorators/withMenuHandler";
 import withLoader from "../../decorators/withLoader";
 import { withDataHandler } from "../../plugins/DocManager/DataHandler";
+import { KEYBINDINGS } from "../../utils/Keybindings";
 import { PLUGINS } from "../../utils/Constants";
 import { ViewPlugin } from "./ViewReactPlugin";
 
@@ -54,8 +55,6 @@ export function withEditorPlugin(ReactComponent, methods = []) {
       updateRightMenu
     } = props;
 
-    const TAB_ACTIVE_EVENT = useMemo(() => `${id}-active`, [id]);
-
     /**
      * Handle submit action on save outdated document
      * @param {string} action : One of options ("cancel", "updateDoc", "overwriteDoc")
@@ -63,7 +62,12 @@ export function withEditorPlugin(ReactComponent, methods = []) {
     const _handleOutdatedSave = React.useCallback(
       action => {
         const getSaveByAction = {
-          updateDoc: () => call("docManager", "reloadDoc", { scope, name }),
+          updateDoc: () =>
+            call(
+              PLUGINS.DOC_MANAGER.NAME,
+              PLUGINS.DOC_MANAGER.CALL.RELOAD_DOC,
+              { scope, name }
+            ),
           overwriteDoc: save
         };
         return action in getSaveByAction ? getSaveByAction[action]() : false;
@@ -84,20 +88,31 @@ export function withEditorPlugin(ReactComponent, methods = []) {
       if (!instance.current.isDirty) return;
 
       if (instance.current.getOutdated()) {
-        call("dialog", "saveOutdatedDocument", {
+        call(PLUGINS.DIALOG.NAME, PLUGINS.DIALOG.CALL.SAVE_OUTDATED_DOC, {
           name,
           scope,
           onSubmit: _handleOutdatedSave
         });
       } else {
         instance.current.getIsNew()
-          ? call("dialog", "newDocument", {
+          ? call(PLUGINS.DIALOG.NAME, PLUGINS.DIALOG.CALL.NEW_DOC, {
               scope,
               onSubmit: newName => save(newName)
             })
           : save();
       }
     }, [call, instance, save, _handleOutdatedSave, scope, name]);
+
+    /**
+     * Save document :
+     *  if document is outdated => prompt alert to the user before saving
+     *  else => Proceed with saving document
+     *    if doc is new => Create document in DB
+     *    else => Update document in DB
+     */
+    const saveAllDocuments = React.useCallback(() => {
+      call(PLUGINS.DOC_MANAGER.NAME, PLUGINS.DOC_MANAGER.CALL.SAVE_DIRTIES);
+    }, [call]);
 
     /**
      * Activate editor : activate editor's keybinds and update right menu
@@ -112,23 +127,29 @@ export function withEditorPlugin(ReactComponent, methods = []) {
      */
     useEffect(() => {
       initRightMenu();
-      addKeyBind("ctrl+s", saveDocument);
-      on(PLUGINS.TABS.NAME, TAB_ACTIVE_EVENT, activateEditor);
+      addKeyBind(KEYBINDINGS.SAVE, saveDocument);
+      addKeyBind(KEYBINDINGS.SAVE_ALL, saveAllDocuments);
+      on(PLUGINS.TABS.NAME, PLUGINS.TABS.ON.ACTIVE_TAB_CHANGE, data => {
+        if (data.id === id) {
+          activateEditor();
+        }
+      });
 
       // Remove key bind on component unmount
       return () => {
-        removeKeyBind("ctrl+s");
-        off(PLUGINS.TABS.NAME, TAB_ACTIVE_EVENT);
+        removeKeyBind(KEYBINDINGS.SAVE);
+        off(PLUGINS.TABS.NAME, PLUGINS.TABS.ON.ACTIVE_TAB_CHANGE);
       };
     }, [
+      id,
       activateEditor,
       addKeyBind,
       removeKeyBind,
-      TAB_ACTIVE_EVENT,
       initRightMenu,
       on,
       off,
-      saveDocument
+      saveDocument,
+      saveAllDocuments
     ]);
 
     return (