+++ src/plugins/views/editors/Callback/dialogs/EditMessage.jsx
@@ -1,7 +1,4 @@
 import React from "react";
-import Loader from "../../_shared/Loader/Loader";
-import MaterialTree from "../../_shared/MaterialTree/MaterialTree";
-import Search from "../../_shared/Search/Search";
 import {
   Typography,
   TextField,
@@ -11,9 +8,13 @@ import {
   DialogActions
 } from "@material-ui/core";
 import { makeStyles } from "@material-ui/core/styles";
-import { searchMessages } from "./utils";
+import { PLUGINS } from "../../../../../utils/Constants";
 import { withTheme } from "../../../../../decorators/withTheme";
 import { DialogTitle } from "../../../../Dialog/components/AppDialog/AppDialog";
+import Loader from "../../_shared/Loader/Loader";
+import MaterialTree from "../../_shared/MaterialTree/MaterialTree";
+import Search from "../../_shared/Search/Search";
+import { searchMessages } from "./utils";
 
 const useStyles = makeStyles(theme => ({
   treeRoot: {
@@ -109,7 +110,11 @@ const EditMessageDialog = props => {
    */
   React.useEffect(() => {
     setLoading(true);
-    call("docManager", "getStore", scope).then(store => {
+    call(
+      PLUGINS.DOC_MANAGER.NAME,
+      PLUGINS.DOC_MANAGER.CALL.GET_STORE,
+      scope
+    ).then(store => {
       store.helper.getAllMessages().then(msgs => {
         if (msgs) _updateMessages(msgs);
         setLoading(false);