+++ src/plugins/views/editors/Flow/Components/Menus/ContainerMenu.jsx
@@ -116,7 +116,10 @@ const ContainerMenu = props => {
     const name = data?.ContainerFlow;
     if (!name) return;
     // Read node template
-    call("docManager", "read", { name, scope: data.model }).then(doc => {
+    call(PLUGINS.DOC_MANAGER.NAME, PLUGINS.DOC_MANAGER.CALL.READ, {
+      name,
+      scope: data.model
+    }).then(doc => {
       setTemplateData(doc.serialize());
     });
   }, [data, call]);