+++ src/plugins/views/editors/Flow/Components/Menus/NodeMenu.jsx
@@ -11,7 +11,7 @@ import {
 } from "@material-ui/core";
 import ExpandLess from "@material-ui/icons/ExpandLess";
 import ExpandMore from "@material-ui/icons/ExpandMore";
-import { DATA_TYPES } from "../../../../../../utils/Constants";
+import { DATA_TYPES, PLUGINS } from "../../../../../../utils/Constants";
 import ParameterEditorDialog from "../../../_shared/KeyValueTable/ParametersEditorDialog";
 import { TABLE_KEYS_NAMES, DIALOG_TITLE } from "../../Constants/constants";
 import MenuDetails from "./sub-components/MenuDetails";
@@ -133,7 +133,10 @@ const NodeMenu = memo(
       const name = data?.Template;
       if (!data?.Template) return;
       // Read node template
-      call("docManager", "read", { name, scope: data.model }).then(doc => {
+      call(PLUGINS.DOC_MANAGER.NAME, PLUGINS.DOC_MANAGER.CALL.READ, {
+        name,
+        scope: data.model
+      }).then(doc => {
         setTemplateData(doc.serialize());
       });
     }, [data, call]);
@@ -176,7 +179,7 @@ const NodeMenu = memo(
           paramType
         };
 
-        const method = "customDialog";
+        const method = PLUGINS.DIALOG.CALL.CUSTOM_DIALOG;
         const args = {
           onSubmit: handleSubmitParameter,
           title: t("Edit {{paramType}}", { paramType }),