+++ src/utils/Constants.js
@@ -1,11 +1,24 @@
-import get from "lodash/get";
-
 export const APP_DEFAULT_CONFIG = "app-ide-ce";
 export const APP_CUSTOM_CONFIG = "app-custom-ide-ce";
 
 export const MANAGER = "manager";
 
-export const VERSION = get(window, "SERVER_DATA.Application.Version", "0.0.1");
+export const APP_INFORMATION = {
+  VERSION: window.SERVER_DATA?.Application?.Version || "0.0.1",
+  LAST_UPDATE: window.SERVER_DATA?.Application?.LastUpdate || "-",
+  CONFIGURATION_FILE: window.SERVER_DATA?.Application?.Configuration || "-",
+  CUSTOM_CONFIGURATION_FILE: window.SERVER_DATA?.CustomConfiguration || "-",
+  DESCRIPTION: window.SERVER_DATA?.Application?.Description || "-"
+};
+
+export const APP_LINKS = {
+  DOCUMENTATION: "https://movai-flow.readme.io/docs",
+  FORUM: "https://discourse.aws.cloud.mov.ai/"
+};
+
+export const BRANDING = {
+  NAME: "Mov.ai Flow™"
+};
 
 export const DISABLED_VALUE = "None";
 
@@ -25,12 +38,50 @@ export const SCOPES = {
   Node: "Node"
 };
 
+export const HOSTS = {
+  TOP_BAR: {
+    NAME: "topBar"
+  },
+  LEFT_PANEL: {
+    NAME: "leftPanel"
+  },
+  MAIN_PANEL: {
+    NAME: "mainPanel"
+  },
+  LEFT_DRAWER: {
+    NAME: "leftDrawer",
+    CALL: {
+      OPEN: "open",
+      CLOSE: "close",
+      TOGGLE: "toggle"
+    }
+  },
+  BOTTOM_BAR: {
+    NAME: "bottomBar"
+  }
+};
+
 export const PLUGINS = {
   DOC_MANAGER: {
     NAME: "docManager",
     CALL: {
       READ: "read",
-      BROADCAST: "broadcast"
+      SAVE: "save",
+      BROADCAST: "broadcast",
+      GET_STORE: "getStore",
+      GET_DOC_FACTORY: "getDocFactory",
+      GET_DOC_TYPES: "getDocTypes",
+      SUBSCRIBE_TO_CHANGES: "subscribeToChanges",
+      UNSUBSCRIBE_TO_CHANGES: "unSubscribeToChanges",
+      GET_DOC_FROM_NAME_TYPE: "getDocFromNameType",
+      CHECK_DOCUMENT_EXISTS: "checkDocumentExists",
+      DISCARD_DOC_CHANGES: "discardDocChanges",
+      RELOAD_DOC: "reloadDoc",
+      COPY: "copy",
+      DELETE: "delete",
+      CREATE: "create",
+      SAVE_ACTIVE_EDITOR: "saveActiveEditor",
+      SAVE_DIRTIES: "saveDirties"
     },
     ON: {
       FLOW_EDITOR: "flowEditor",
@@ -41,7 +92,20 @@ export const PLUGINS = {
       SAVE_DOC: "saveDoc"
     }
   },
-  TABS: { NAME: "tabs", CALL: { OPEN_EDITOR: "openEditor" } },
+  TABS: {
+    NAME: "tabs",
+    CALL: {
+      OPEN: "open",
+      CLOSE: "close",
+      OPEN_EDITOR: "openEditor",
+      UPDATE_TAB_ID: "updateTabId",
+      GET_ACTIVE_TAB: "getActiveTab"
+    },
+    ON: {
+      OPEN_EDITOR: "openEditor",
+      ACTIVE_TAB_CHANGE: "activeTabChange"
+    }
+  },
   RIGHT_DRAWER: {
     NAME: "rightDrawer",
     CALL: {
@@ -58,16 +122,39 @@ export const PLUGINS = {
     NAME: "dialog",
     CALL: {
       NEW_DOC: "newDocument",
+      SAVE_OUTDATED_DOC: "saveOutdatedDocument",
       CONFIRMATION: "confirmation",
+      COPY_DOC: "copyDocument",
       FORM_DIALOG: "formDialog",
+      SELECT_SCOPE_MODAL: "selectScopeModal",
+      CLOSE_DIRTY_DOC: "closeDirtyDocument",
+      CUSTOM: "custom",
       CUSTOM_DIALOG: "customDialog",
       ALERT: "alert"
     }
+  },
+  ALERT: {
+    NAME: "alert",
+    CALL: {
+      SHOW: "show"
+    }
+  },
+  MAIN_MENU: {
+    NAME: "mainMenu"
+  },
+  EXPLORER: {
+    NAME: "explorer"
+  },
+  PLACEHOLDER: {
+    NAME: "placeholder"
   },
   FLOW_EXPLORER: {
     NAME: "FlowExplorer",
     CALL: {},
     ON: { ADD_NODE: "addNode" }
+  },
+  SYSTEM_BAR: {
+    NAME: "systemBar"
   }
 };