+++ src/utils/Utils.js
@@ -83,12 +83,47 @@ export const stopPropagation = e => {
 };
 
 /**
+ * Returns the document version from an URL
+ * @param {String} url
+ * @returns {String}
+ */
+export function getVersionFromUrl(url) {
+  if (!url) return "";
+  const splittedUrl = url.split("/");
+  return splittedUrl[3];
+}
+
+/**
  * Returns the document name from an URL
  * @param {String} url
  * @returns {String}
  */
 export function getNameFromURL(url) {
-  return url?.substring(url.lastIndexOf("/") + 1);
+  if (!url) return "";
+  const splittedUrl = url.split("/");
+  return splittedUrl[2];
+}
+
+/**
+ * Returns the document scope from an URL
+ * @param {String} url
+ * @returns {String}
+ */
+export function getScopeFromURL(url) {
+  if (!url) return "";
+  const splittedUrl = url.split("/");
+  return splittedUrl[1];
+}
+
+/**
+ * Returns the document workspace from an URL
+ * @param {String} url
+ * @returns {String}
+ */
+export function getWorkspaceFromUrl(url) {
+  if (!url) return "";
+  const splittedUrl = url.split("/");
+  return splittedUrl[0];
 }
 
 /**
@@ -159,3 +194,20 @@ export const getHomeTab = () => {
     content: viewPlugin.render()
   });
 };
+
+/**
+ * Trigger a simulated mouse click (react element)
+ * @param {*} element
+ */
+export function simulateMouseClick(element) {
+  ["mousedown", "click", "mouseup"].forEach(mouseEventType =>
+    element.dispatchEvent(
+      new MouseEvent(mouseEventType, {
+        view: window,
+        bubbles: true,
+        cancelable: true,
+        buttons: 1
+      })
+    )
+  );
+}